config/requirements.txt
# general

azure-identity==1.21.0
azure-appconfiguration==1.7.1

# app container apps

azure-mgmt-appcontainers==3.2.0

# aifoundry

azure-ai-projects==1.0.0b11  
azure-mgmt-cognitiveservices==13.6.0

# search

requests==2.32.3
azure-keyvault-secrets==4.9.0
